Added: Loading of wcs_domole
Added some code that was missing. 
The eventscripts file containing the command wcs_domole should now also be loaded when the WCS Essentials are being loaded, as initially intended. 
With this update all the races using the mole skill will now be fully functional and working.


Added a loading of the wcs_domole 

Added a loading of the wcs_domole