Added: Emulator Run.Py
Added the Run.py file to the Warcraft-Source Essentials. This file overwrites the already existing run.py within the Eventscripts Emulator.
By replacing this run.py file with the already existing run.py file in the emulator a lot of bugs and issues related to es_delayed and wcs commands will be fixed. This implementation fixes hundred or more race related bugs and issues, without exaggeration, so it is very important that you get this update.