Added: Mole Collision Fix
If a player becomes a mole he would often find himself stuck in an enemy player once the freeze time expired.

With the implementation of this fix all playrs will have No-Block enabled for the first 3 seconds after the freeze time period. This only happens when a player becomes a mole.