Added: Fixes For Counter-Strike: Source
Added a new script to the Warcraft-Source Essentials package.

The script will only be loaded if you are running a Counter-Strike: Source server, as the problems are not present on CS:GO.

The script allows the usage of es damage to apply the intended damage to the targeted player.

The script also fixes the issue where gravity alteration on one race would carry over to the next race used, despite the next race not having any gravity skills themselves.

When running this script you should NOT be running the old firefix.py since this new and improved version is meant to replace it completely.